refactor: extract RulesFacade and SimulatorFacade from CLIFacade
Split rules CRUD and simulator logic into dedicated facades for better
separation of concerns. CLIFacade retains collections, packs, service,
and system operations.
