+++ Sources/KeyPathCLI/Commands/Collection/CollectionEnableCommand.swift
@@ -47,6 +47,6 @@ struct CollectionEnable: AsyncParsableCommand {
             throw error.code.exitCode
         }
 
-        try await applyConfigurationOrHint(facade: facade, apply: apply, context: ctx)
+        try await applyConfigurationOrHint(apply: apply, context: ctx)
     }
 }
