+++ Sources/KeyPathCLI/Commands/Pack/PackInstallCommand.swift
@@ -87,7 +87,7 @@ struct PackInstall: AsyncParsableCommand {
             }
 
             if result.action == "installed" {
-                try await applyConfigurationOrHint(facade: facade, apply: apply, context: ctx)
+                try await applyConfigurationOrHint(apply: apply, context: ctx)
             }
         } catch let notFound as CLIPackNotFound {
             spinner.fail("Pack not found: '\(notFound.query)'")
